requirements.txt
Django==1.6.5
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
topia.termextract==1.1.0
beautifulsoup4==4.3.2
gunicorn==19.0.0
static3==0.5.1
psycopg2==2.5.3
wsgiref==0.1.2
mandrill>=1.0.6
South>=1.0.2
